* Re-instated case conversion but this time to "Capitalisation Of Initials" for consistant appearance across l10n tabs.
The previous commit removed the case conversion on the titles -- that was a mistake as I assumed all the translations would have "instantly" followed the changes the dev team have made to the English. Some other languages are still using all lower-case tab labels. This meant that in French for example, some of the l10n tab titles were "Like This" and others "like this". Very messy.